Former-commit-id: e6d46e00b4bf6b208252c6675827a4cc88cb22f7

+++ Tools/ResourceEditor/Classes/Qt/Scene/System/LandscapeEditorSystem.h
@@ -27,7 +27,6 @@
 =====================================================================================*/
 
 
-
 #ifndef __LANDSCAPE_EDITOR_SYSTEM__
 #define __LANDSCAPE_EDITOR_SYSTEM__
 
@@ -67,7 +66,7 @@ protected:
     
     
     DAVA::Texture* cursorTexture;
-    DAVA::uint32 cursorSize;
+    DAVA::float32 cursorSize;
 	Vector2 cursorPosition;
     Vector2 prevCursorPos;
